refactor: move the api_docs to the openapi module
