+++ Dockerfile
@@ -35,7 +35,7 @@ FROM rust:latest
 
 COPY --from=builder /usr/local/cargo/bin/myc-api /usr/local/bin/myc-api
 
-COPY ports/api/src/api_docs/redoc.config.json /home/redoc.config.json
+COPY ports/api/src/openapi/redoc.config.json /home/redoc.config.json
 COPY templates /home/templates
 
 ENV UTOIPA_REDOC_CONFIG_FILE=/home/redoc.config.json
